Add FileDownloadChunk callback called after sending ack
The current pnAuthClient::onFileDownloadChunk callback is called before sending the acknowledge message to the server. This callback is typically used to send the next command to the server. Some servers (notably MOSS) don’t like it if you request a new file before acknowledging receipt of the previous one. To handle that, the callback should be called after sending the ack message. To avoid disrupting functionality of existing library users, the existing onFileDownloadChunk method is left alone and a new additional method pnAuthClient::onFileDownloadChunkPostAck is introduced.