Add 64-bit integer PRC Param writer
Pull request #107 changed the way the maximum number of emitters was
stored such that it now has a type of size_t. On x64, this is a 64-bit
integer for which there was no 64-bit PRC overload. This adds a 64-bit
overload in the style of the already existing ones.