Pack DSpan GeoSpans into single vertex buffer.
This changeset optimizes the output of
`plDrawableSpans::composeGeometry` by attempting to pack spans into a
one vertex buffer in plGBufferGroup. These vertex buffers correspond to
DirectX vertex buffers in Plasma. There are many buffer group formats in
Plasma, and the code attempts to not obliterate anything another user of
the buffer group has done.

This also fixes an issue in which the index data of a `plGeometrySpan`
could be lost if the number of indices changed after being initialized
or set by IO.