Fix round-tripping plPythonFileMod through HSPlasma.
This fixes an issue in which PythonFileMods with string parameters would
be shattered by round-tripping through HSPlasma. The spurious null
terminator was being accumulated into the parameter, causing two nulls
to be serialized. This caused a null byte to be passed to CWE's Python,
resulting in a difficult to debug scenario where we had a seemingly
valid string object in Python but the validation in `PyArg_ParseTuple`
failed.