Sort and fixup keyrings on read.
Cyan PRPs do not necessarily have their keyrings in order by object ID,
and some object IDs may actually be skipped. The former tends to be seen
primarily in Texture PRPs and the latter in PRPs processed by
plPageOptimizer (see: Kadish_District_opt-*.prp).

Before this change and the log message were introduced, you could see
this by opening a whole MOUL age in PRPShop, saving all the pages but
taking care to save the Textures page *last*. In an *external* MOUL
client - the Textures would be mis-linked due to the object IDs being
changed on write of the Textures PRP.

This adds a forced re-ID pass on read of a page to ensure that IDs are
sane before they are written. This ensures that page reading and writing
are order independent. You still need to operate on a full Age when
dealing with MOUL/EOA files, however.