Fix incorrect keyring flag.
Legacy MOUL clients expect for the keyring to be sorted by name. If keys
are not sorted and the flag is not applied, then searches by name will
silently fail. This was observed when someone attempted to add a new
image to GUI_District_BkBookImages.prp and use it a journal. Due to this
issue, no image was returned by the journal's search.