Add ResMgr option to preserve Object IDs
After several attempts to address issues around Object ID mismatches,
we've adopted a stance of libHSPlasma reassigning all Object IDs on
read. This generally works okay (with caveats around cross-page
references) but causes problems if you're wanting to round-trip data
through libHSPlasma without modifying it.

This adds a `preserveObjIDs` parameter to plResManager that tells it
not to change object IDs when reading. This is **not safe** if you are
adding or removing any objects from the PRP file, but should be safe if
you're only wanting to edit some object data.

This also sets up the `prcdc` tool to always dump out the original
Object IDs to get an accurate view of the data.

When running `prcc` to compile from PRC files, there's a new
`--keep-ids` option to use the object ID values as read from the PRC.
This enables us to round-trip through `prcdc/prcc` without renumbering
all the data, which should help us to track down any PRC parsing issues.

Co-Authored-By: Michael Hansen <zrax0111@gmail.com>