Add generic typing for PyHSPlasma's `plKey`.
This allows you to declare what type of `hsKeyedObject` subclass the
`plKey` is pointing to for much improved type hints. I've further gone
through and annotated the majority of the existing hints to indicate
which `hsKeyedObject` subclasses are held. In some cases, Plasma itself
doesn't actually dereference the keys but rather uses the keys as
message receivers, so I did not bother with adding type hints for these.
A future PR might add more `TypeVar`s for classes that accept, for
example, `plAnimCmdMsg` and propagate that common case into the type
hints.