Fix erroneous `TypeError` from all bindings using the `s#` code.
As of Python 3.10, all argument codes with the `#` suffix are disabled
by default. To re-enable them, we must define `PY_SSIZE_T_CLEAN` and
change all of the types from `int` to `Py_ssize_t`. In Python 3.9 and
lower, `#` arguments expect the `int` data type by default. The macro is
backwards compatible.