Avoid WORDS_BIGENDIAN conflict with Python
Python uses the same macro name internally, and this results in compiler
warnings about redefining it. Since we're updating this, we might as
well move to using CMake's endianness detection as opposed to relying on
macros being defined (or not defined) in system headers because that has
proven unreliable in the past.