Re-implement previous plLocation comparison behavior for FindPage, etc.
Otherwise the plAgeInfo methods getPageLoc, getCommonPageLoc, and
getPageLocs can no longer be used, because they return locations whose
flags do not necessarily match the flags of the real corresponding pages
(and those methods *cannot* be fixed to return the correct flags in the
general case, because of the kItinerant flag, which cannot be predicted
from other properties of the page).

This fixes the crash when opening an .age file in PrpShop, caused by the
preceding commits.