Fix plMipmap JPEG decompression using the wrong channel for JPEG alpha
JPEG alpha images store the alpha data in the red channel, but our
decompression code incorrectly extracted the blue channel instead,
leading to corrupted alpha channels for mipmaps with JPEG alpha.

The uncompressed buffers in question are stored in BGRA order in memory,
but the channel extraction/swapping code accesses them as an array of
32-bit ints. From this perspective, on little-endian systems, the buffer
"looks like" ARGB. So to use the red channel as alpha, we need to shift
each 32-bit int 8 bits to the left and mask out the low 24 bits.